Add a new alias for vim
The terminal used to froze while using vim when <C-S> was pressed. In
order to prevent that from happening I created an alias which executes
an extra command each time you run vim which is the one that solves this
issue. Aside of that, I also needed the shortcut <C-S> for inserting
eruby tags, so this way it is now usable.